Clean up the framework
