+++ framework/Cargo.toml
@@ -11,7 +11,7 @@ chacha20poly1305 = "0"
 chrono = { version = "0", default-features = false, features = ["std"] }
 dashmap = "4"
 framework-derive = { path = "./framework-derive" }
-futures = "0"
+futures-util = "0"
 itertools = "0"
 patreon = { branch = "main", git = "https://github.com/AsianIntel/patreon" }
 roblox = { path = "../roblox" }
