+++ framework/src/context.rs
@@ -1,6 +1,5 @@
 use chacha20poly1305::ChaCha20Poly1305;
 use dashmap::{DashMap, DashSet};
-use futures::{Future, FutureExt};
 use itertools::Itertools;
 use patreon::Client as Patreon;
 use roblox::Client as Roblox;
@@ -18,17 +17,11 @@ use std::{
     borrow::ToOwned,
     collections::{HashMap, HashSet},
     ops::Deref,
-    pin::Pin,
     sync::Arc,
-    task::{Context, Poll},
 };
 use twilight_gateway::Cluster;
-use twilight_http::{
-    request::{application::UpdateOriginalResponse, prelude::CreateMessage},
-    Client as Http, Error as DiscordHttpError,
-};
+use twilight_http::Client as Http;
 use twilight_model::{
-    application::component::Component,
     channel::embed::Embed,
     id::{ChannelId, GuildId, InteractionId, MessageId, RoleId, UserId, WebhookId},
     user::User,
@@ -36,7 +29,10 @@ use twilight_model::{
 use twilight_standby::Standby;
 use twilight_util::link::webhook;
 
-use crate::error::{CommandError, RoError};
+use crate::{
+    error::{CommandError, RoError},
+    respond::Responder,
+};
 
 pub struct BotContextRef {
     // Config Items
@@ -470,83 +466,6 @@ impl BotContext {
                 .embed(embed)
                 .unwrap()
                 .await;
-        }
-    }
-}
-
-pub struct Responder<'a> {
-    message: Option<CreateMessage<'a>>,
-    interaction: Option<UpdateOriginalResponse<'a>>,
-}
-
-impl<'a> Responder<'a> {
-    pub fn new(ctx: &'a CommandContext) -> Self {
-        ctx.interaction_token.as_ref().map_or_else(
-            || Self {
-                message: Some(ctx.bot.http.create_message(ctx.channel_id)),
-                interaction: None,
-            },
-            |interaction_token| Self {
-                message: None,
-                interaction: Some(
-                    ctx.bot
-                        .http
-                        .update_interaction_original(interaction_token)
-                        .unwrap(),
-                ),
-            },
-        )
-    }
-
-    pub fn content(mut self, content: impl Into<String>) -> Self {
-        let content = content.into();
-        if let Some(interaction) = self.interaction {
-            self.interaction = Some(interaction.content(Some(content)).unwrap());
-        } else if let Some(message) = self.message {
-            self.message = Some(message.content(content).unwrap());
-        }
-        self
-    }
-
-    pub fn component(mut self, component: Component) -> Self {
-        if let Some(interaction) = self.interaction {
-            self.interaction = Some(interaction.component(component).unwrap());
-        } else if let Some(message) = self.message {
-            self.message = Some(message.component(component).unwrap());
-        }
-        self
-    }
-
-    pub fn components(mut self, components: Vec<Component>) -> Self {
-        if let Some(interaction) = self.interaction {
-            self.interaction = Some(interaction.components(components).unwrap());
-        } else if let Some(message) = self.message {
-            self.message = Some(message.components(components).unwrap());
-        }
-        self
-    }
-
-    pub fn embed(mut self, embed: Embed) -> Self {
-        if let Some(interaction) = self.interaction {
-            self.interaction = Some(interaction.embeds(Some(vec![embed])).unwrap());
-        } else if let Some(message) = self.message {
-            self.message = Some(message.embed(embed).unwrap());
-        }
-        self
-    }
-}
-
-#[allow(clippy::option_if_let_else)]
-impl Future for Responder<'_> {
-    type Output = Result<(), DiscordHttpError>;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        if let Some(interaction) = self.interaction.as_mut() {
-            interaction.poll_unpin(cx)
-        } else if let Some(message) = self.message.as_mut() {
-            message.poll_unpin(cx).map(|p| p.map(|_| ()))
-        } else {
-            Poll::Ready(Ok(()))
         }
     }
 }
