+++ framework/src/extensions/mod.rs
@@ -1,0 +1,3 @@
+mod embed;
+
+pub use embed::EmbedExtensions;
