+++ framework/src/prelude.rs
@@ -1,13 +1,13 @@
 pub use crate::arguments::*;
-pub use crate::command::Command;
+pub use crate::command::{Command, CommandResult};
 pub use crate::context::CommandContext;
 pub use crate::error::*;
+pub use crate::extensions::*;
 pub use crate::utils::*;
-pub use crate::CommandResult;
 
 pub use framework_derive::FromArgs;
 pub use tower::{Service, ServiceExt};
-pub use twilight_embed_builder::{EmbedBuilder, EmbedFieldBuilder};
+pub use twilight_embed_builder::*;
 pub use twilight_model::application::{
     component::{
         action_row::ActionRow,
